Add a tool to compile all snippets
This tool will only compile all the snippets in one go. It will not
generate code or link. This is only to be used as a CI job. It's
significantly faster to only compile and compile all snippets at once
then to build each snippet individually. Since the snippets are not
linked there is not problem with duplicated symbols, like `main`.