[base] Fix SafeUtf unittests
The primary fix was by adding the opEquals overload to
UTF8shift, but an alterative would be appending .val
onto the end of anything that returns an UTF8shift.

The separation between "iter" and "idx" is required
because you can't implicitly convert from UTF8index to
a size_t, nor vice versa.