Update Converter.d to 4.7.3
This also fixes an issue where LDC would fail to compile due to a
function signature mismatch with the g_utf16_to_utf8 declared in the
OS module (an alias to glib_object.d).  They are still declared here due
to wchar being easier to work with than ushort when dealing with
characters.