change ImageLoader import to a .all
This change is required to allow a call to Object.factory to work
as intended.

For such calls to work, D needs to know about the module (i.e.
the symbol needs to be resolvable). This specific change is in
relation to a change that will happen in the FileFormat class
which will create a child of FileFormat dynamically.